Refactored more types
