+++ packages/shared/src/types/OpenAI.ts
@@ -1,0 +1,4 @@
+export interface OpenAIError extends Error {
+  status?: number;
+  message: string;
+}
