+++ packages/shared/src/types/index.ts
@@ -1,1 +1,3 @@
 export * from "./Story";
+export * from "./OpenAI";
+export * from "./Generation";
